requirements.txt
acme==2.6.0
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certbot==2.6.0
certbot-nginx==2.6.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.2.0
chromedrivermanager==0.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
configobj==5.0.8
configparser==6.0.0
contourpy==1.0.7
coverage==7.2.3
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
Django==4.2.3
django-admin-logs==1.0.2
django-apscheduler==0.6.2
django-chartjs==2.3.0
django-cors-headers==4.1.0
django-csp==3.7
django-log-viewer==1.1.7
django-mail-admin==0.3.2
django-ratelimit==4.1.0
django-rest-framework-async==0.1.0
django-sslserver==0.22
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-firebase==0.3.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.11.0
firebase-admin==2.18.0
flake8==6.0.0
Flask==2.3.2
fonttools==4.39.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==1.34.0
google-api-python-client==2.101.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.48.2
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.13.1
ipywidgets==8.0.6
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
josepy==1.13.0
jsonfield==3.1.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.6
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nest-asyncio==1.5.6
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.1.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==2.0.1
parsedatetime==2.6
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyOpenSSL==23.1.1
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
sourcery==1.2.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tiktoken==0.3.3
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webdriver-manager==3.8.6
websocket-client==1.5.1
Werkzeug==2.3.4
wget==3.2
widgetsnbextension==4.0.7
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.8.2
